requirements.txt
python-telegram-bot==20.7
httpx~=0.25.2
aio-pika==9.4.3
python-dotenv==1.0.1
